requirements.txt
python-telegram-bot
google-generativeai
psycopg2-binary
boto3
python-dotenv
Pillow
SpeechRecognition
